finalize

+++ src/main/java/com/flashcards/repository/FlashcardListRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 import com.flashcards.model.FlashcardList;
 
-public interface FlashcardListRepository extends JpaRepository <FlashcardList, Integer>{
+public interface FlashcardListRepository extends JpaRepository <FlashcardList, Long>{
 	FlashcardList findByName (String name);
 }
